Update tests 1900 after review
